+++ st2actions/st2actions/config.py
@@ -69,11 +69,23 @@ workflow_opts = [
 ]
 CONF.register_opts(workflow_opts, group='workflow')
 
+mistral_opts = [
+    cfg.StrOpt('v2_base_url', default='http://localhost:8989/v2/',
+               help='Mistral v2 API server root endpoint.'),
+]
+CONF.register_opts(mistral_opts, group='mistral')
+
 history_opts = [
     cfg.StrOpt('logging', default='conf/logging.history.conf',
                help='Location of the logging configuration file.')
 ]
 CONF.register_opts(history_opts, group='history')
+
+resultstracker_opts = [
+    cfg.StrOpt('logging', default='conf/logging.resultstracker.conf',
+               help='Location of the logging configuration file.')
+]
+CONF.register_opts(resultstracker_opts, group='resultstracker')
 
 
 def parse_args(args=None):