+++ st2actions/tests/unit/test_runner_container_service.py
@@ -13,11 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-try:
-    import simplejson as json
-except ImportError:
-    import json
-
 import os
 
 from oslo.config import cfg
@@ -95,10 +90,3 @@ class RunnerContainerServiceTest(unittest2.TestCase):
         expected_path = os.path.join('/tmp', ACTION_LIBS_DIR)
         self.assertEqual(acutal_path, expected_path, 'Action libs path doesn\'t match.')
         cfg.CONF.content.packs_base_path = orig_path
-
-    def test_report_result_json(self):
-        service = RunnerContainerService()
-        result = '["foo", {"bar": ["baz", null, 1.0, 2]}]'
-        service.report_result(result)
-        self.assertEqual(json.dumps(service.get_result()), result,
-                         'JON results aren\'t handled right')