+++ st2common/st2common/models/api/action.py
@@ -21,6 +21,7 @@ from st2common import log as logging
 from st2common.models.api.base import BaseAPI
 from st2common.models.api.tag import TagsHelper
 from st2common.models.db.action import (RunnerTypeDB, ActionDB, ActionExecutionDB)
+from st2common.models.db.action import ActionExecutionStateDB
 from st2common.constants.action import ACTIONEXEC_STATUSES
 
 
@@ -68,6 +69,12 @@ class RunnerTypeAPI(BaseAPI):
                 "type": "string",
                 "required": True
             },
+            "query_module": {
+                "description": "The python module that implements the "
+                               "results tracker (querier) for the runner.",
+                "type": "string",
+                "required": False
+            },
             "runner_parameters": {
                 "description": "Input parameters for the action runner.",
                 "type": "object",
@@ -98,6 +105,8 @@ class RunnerTypeAPI(BaseAPI):
         model = super(cls, cls).to_model(runnertype)
         model.enabled = bool(runnertype.enabled)
         model.runner_module = str(runnertype.runner_module)
+        if getattr(runnertype, 'query_module', None):
+            model.query_module = str(runnertype.query_module)
         model.runner_parameters = getattr(runnertype, 'runner_parameters', dict())
         return model
 
@@ -280,4 +289,47 @@ class ActionExecutionAPI(BaseAPI):
         model.context = getattr(execution, 'context', dict())
         model.callback = getattr(execution, 'callback', dict())
         model.result = getattr(execution, 'result', None)
+        return model
+
+
+class ActionExecutionStateAPI(BaseAPI):
+    """
+    System entity that represents state of an action in the system.
+    This is used only in tests for now.
+    """
+    model = ActionExecutionStateDB
+    schema = {
+        "title": "ActionExecutionState",
+        "description": "Execution state of an action.",
+        "type": "object",
+        "properties": {
+            "id": {
+                "description": "The unique identifier for the action execution state.",
+                "type": "string"
+            },
+            "execution_id": {
+                "type": "string",
+                "description": "ID of the action execution.",
+                "required": True
+            },
+            "query_context": {
+                "type": "object",
+                "description": "query context to be used by querier.",
+                "required": True
+            },
+            "query_module": {
+                "type": "string",
+                "description": "Name of the query module.",
+                "required": True
+            }
+        },
+        "additionalProperties": False
+    }
+
+    @classmethod
+    def to_model(cls, state):
+        model = super(cls, cls).to_model(state)
+        model.query_module = state.query_module
+        model.execution_id = state.execution_id
+        model.query_context = state.query_context
         return model